collection/requirements.txt
backcall==0.1.0
beautifulsoup4==4.9.1
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Deprecated==1.2.10
Genderize==0.3.1
googlemaps==4.4.0
idna==2.9
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
numpy==1.18.4
pandas==1.0.3
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
PyGithub==1.51
Pygments==2.6.1
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
pyxDamerauLevenshtein==1.6
requests==2.23.0
six==1.14.0
soupsieve==2.0.1
traitlets==4.3.3
urllib3==1.25.9
wcwidth==0.1.9
wrapt==1.12.1
